Move CAS2v2 report models into service
