+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/cas2v2/service/Cas2v2ReportsService.kt
@@ -4,14 +4,16 @@ import org.apache.poi.ss.usermodel.WorkbookFactory
 import org.jetbrains.kotlinx.dataframe.api.toDataFrame
 import org.jetbrains.kotlinx.dataframe.io.writeExcel
 import org.springframework.stereotype.Service
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ApplicationOrigin
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2.jpa.entity.Cas2ApplicationStatusUpdatesReportRepository
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2.jpa.entity.Cas2SubmittedApplicationReportRepository
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2.jpa.entity.Cas2UnsubmittedApplicationsReportRepository
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2.model.Cas2ServiceOrigin
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2v2.reporting.model.ApplicationStatusUpdatesReportRow
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2v2.reporting.model.SubmittedApplicationReportRow
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2v2.reporting.model.UnsubmittedApplicationsReportRow
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2v2.service.ApplicationStatusUpdatesReportRow
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2v2.service.SubmittedApplicationReportRow
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2v2.service.UnsubmittedApplicationsReportRow
 import java.io.OutputStream
+import java.time.LocalDate
 
 @Service
 class Cas2v2ReportsService(
@@ -76,3 +78,41 @@ inline fun <reified T> Iterable<T>.toCas2v2Report(outputStream: OutputStream) = 
     outputStream = outputStream,
     factory = WorkbookFactory.create(true),
   )
+
+@Suppress("LongParameterList")
+class ApplicationStatusUpdatesReportRow(
+  val eventId: String,
+  val applicationId: String,
+  val personCrn: String,
+  val personNoms: String,
+  val newStatus: String,
+  val updatedAt: String,
+  val updatedBy: String,
+  val statusDetails: String,
+  val applicationOrigin: String,
+)
+
+data class SubmittedApplicationReportRow(
+  val eventId: String,
+  val applicationId: String,
+  val personCrn: String,
+  val personNoms: String?,
+  val referringPrisonCode: String?,
+  val preferredAreas: String?,
+  val hdcEligibilityDate: LocalDate?,
+  val conditionalReleaseDate: LocalDate?,
+  val submittedAt: String,
+  val submittedBy: String,
+  val startedAt: String,
+  val applicationOrigin: ApplicationOrigin,
+  val bailHearingDate: LocalDate?,
+)
+
+class UnsubmittedApplicationsReportRow(
+  val applicationId: String,
+  val personCrn: String,
+  val personNoms: String?,
+  val startedAt: String,
+  val startedBy: String,
+  val applicationOrigin: ApplicationOrigin,
+)
